This is a temporary fix until we can definitively prove what the problem with the tests are.
There appears to be some sort of issue where either variable scope is not what was expected or a third party library needs cleared between test groups.
