+++ .eslintrc.json
@@ -1,7 +1,7 @@
 {
-  "env": {
-    "commonjs": true,
-    "es6": true,
-    "node": true
-  }
+    "extends": "standard",
+    "installedESLint": true,
+    "plugins": [
+        "standard"
+    ]
 }
