+++ package.json
@@ -57,5 +57,10 @@
     "supertest-as-promised": "^1.0.0",
     "uuid": "^9.0.0",
     "validator": "^5.7.0"
+  },
+  "devDependencies": {
+    "eslint": "^8.50.0",
+    "eslint-config-standard": "^17.1.0",
+    "eslint-plugin-standard": "^5.0.0"
   }
 }
